pyproject.toml
[project]
name = "simpler_env"
version = "0.0.1"
description = "SimplerEnv"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "mani_skill2_real2sim",
    "numpy<2.0",
    "mediapy",
    "transforms3d",
    "matplotlib",
    "statsmodels",
]


[project.optional-dependencies]
torch = [
    "tensorflow-cpu",
]

[tool.black]
# https://github.com/psf/black
line-length = 120
target-version = ["py310"]
exclude = "(.eggs|.git|.hg|.mypy_cache|.nox|.tox|.venv|.svn|_build|buck-out|build|dist)"

[tool.isort]
profile = "black"
line_length = 120
force_sort_within_sections = "True"
order_by_type = "False"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
mani_skill2_real2sim = { path = "ManiSkill2_real2sim", editable = true }